[frontend] reorganize
