+++ opencti-platform/opencti-front/src/private/components/pir/PirHeader.tsx
@@ -18,11 +18,11 @@ import React, { useState } from 'react';
 import { Box, Button, Typography } from '@mui/material';
 import FormAuthorizedMembersDialog from '@components/common/form/FormAuthorizedMembersDialog';
 import PirPopover from './PirPopover';
-import PirEdition from './PirEdition';
+import PirEdition from './pir_form/PirEdition';
 import { PirHeaderFragment$key } from './__generated__/PirHeaderFragment.graphql';
 import { useFormatter } from '../../../components/i18n';
 import Breadcrumbs from '../../../components/Breadcrumbs';
-import { PirEditionFragment$key } from './__generated__/PirEditionFragment.graphql';
+import { PirEditionFragment$key } from './pir_form/__generated__/PirEditionFragment.graphql';
 import { authorizedMembersToOptions, useGetCurrentUserAccessRight } from '../../../utils/authorizedMembers';
 import { PIRAPI_PIRUPDATE, SETTINGS_SETACCESSES } from '../../../utils/hooks/useGranted';
 import Security from '../../../utils/Security';
