+++ opencti-platform/opencti-front/src/private/components/pir/pir_form/PirCreationFormStepper.tsx
@@ -12,9 +12,10 @@ Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 */
+
 import { Step, StepButton, Stepper } from '@mui/material';
 import React from 'react';
-import { useFormatter } from '../../../components/i18n';
+import { useFormatter } from '../../../../components/i18n';
 
 interface PirCreationFormStepperProps {
   step: number
