+++ opencti-platform/opencti-front/src/private/components/pir/pir_form/PirEditionForm.tsx
@@ -17,11 +17,11 @@ import * as Yup from 'yup';
 import { Field, Form, Formik } from 'formik';
 import React from 'react';
 import { PirEditionFormData } from './pir-form-utils';
-import { useFormatter } from '../../../components/i18n';
+import { useFormatter } from '../../../../components/i18n';
 import { PirEditionFragment$data } from './__generated__/PirEditionFragment.graphql';
-import TextField from '../../../components/TextField';
-import MarkdownField from '../../../components/fields/MarkdownField';
-import { fieldSpacingContainerStyle } from '../../../utils/field';
+import TextField from '../../../../components/TextField';
+import MarkdownField from '../../../../components/fields/MarkdownField';
+import { fieldSpacingContainerStyle } from '../../../../utils/field';
 
 export type PirEditionFormInputKeys = keyof PirEditionFormData;
